+++ modi/member-service/src/main/java/com/coc/modi/member/domain/MemberRepository.java
@@ -1,0 +1,14 @@
+package com.coc.modi.member.domain;
+
+import java.util.Optional;
+
+public interface MemberRepository {
+
+    boolean existsByEmail(String email);
+
+    Optional<Member> findByEmail(String email);
+
+    Member save(Member member);
+
+    Optional<Member> findById(Long memberId);
+}
