+++ modi/member-service/src/main/java/com/coc/modi/member/infrastructure/MemberRepositoryAdapter.java
@@ -1,6 +1,7 @@
 package com.coc.modi.member.infrastructure;
 
 import com.coc.modi.member.domain.Member;
+import com.coc.modi.member.domain.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -8,27 +9,27 @@ import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
-public class MemberRepository {
+public class MemberRepositoryAdapter implements MemberRepository {
 
     private final MemberJpaRepository memberJpaRepository;
 
+    @Override
     public boolean existsByEmail(String email) {
-
         return memberJpaRepository.existsByEmail(email);
     }
 
+    @Override
     public Optional<Member> findByEmail(String email) {
-
         return memberJpaRepository.findByEmail(email);
     }
 
+    @Override
     public Member save(Member member) {
-
         return memberJpaRepository.save(member);
     }
 
+    @Override
     public Optional<Member> findById(Long memberId) {
-
         return memberJpaRepository.findById(memberId);
     }
 }
